changed 'pow' to 'powf' to prevent '6 methods overload error'

+++ plugins/environment/mystarbox/star.cpp
@@ -104,7 +104,7 @@ void Star::DrawStar3D(iGraphics3D* g3d, const iCamera* c)
   float scale = 0;
   if (apr_lum <= 6.5)
   {
-    scale = pow(SB_APR_MAG_EXP, (apr_lum > 0.0f ? apr_lum : -1 * apr_lum));
+    scale = powf(SB_APR_MAG_EXP, (apr_lum > 0.0f ? apr_lum : -1 * apr_lum));
   }
   int offset = static_cast<int>(scale / 2);
   int img_size;
